requirements.dev.txt
apache-beam[gcp]==2.50.0
hydra-core==1.3.2
python-dotenv
google-cloud-pubsub